+++ WebApi/Models/WebCategory.cs
@@ -4,9 +4,9 @@ namespace WebApi.Models
 {
     public class WebCategory
     {
-        public BigInteger Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
-        public BigInteger? ParentId { get; set; }
+        public long? ParentId { get; set; }
         public string Description { get; set; }
         public string IconPath { get; set; }
     }
