+++ WebApi/Models/WebProduct.cs
@@ -9,7 +9,7 @@ namespace WebApi.Models
 {
     public class WebProduct
     {
-        public BigInteger Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Slug { get; set; }
         public DateTime DateCreated { get; set; }
@@ -29,7 +29,7 @@ namespace WebApi.Models
     }
     public class Attribute
     {
-        public BigInteger Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Value { get; set; }
     }
